### 2023-02-15 #5152 @editorialbot
[PRE REVIEW]: spINAR: An R Package for Semiparametric and Parametric Estimation and Bootstrapping of Integer-Valued Autoregressive (INAR) Models
**Submitting author:** @MFaymon (<a href="http://orcid.org/0000-0001-5565-0711">Maxime  Faymonville</a>)
**Repository:** https://github.com/MFaymon/spINAR
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
