### 2023-02-28 #5152 @MFaymon
I suggest Manuel Stapper (ManuelStapper) as reviewer for our paper who already agreed to the review. Additionally, I suggest Dom-Owens-UoB, mehmet-ozcan and SaranjeetKaur from the list of people who have already agreed to review for JOSS.
