+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: spINAR
 Type: Package
 Title: Semiparametric and Parametric Estimation and Bootstrapping of Integer-Valued Autoregressive (INAR) Models
 Version: 0.1.0
-Date: 2023-02-07
+Date: 2023-03-22
 Authors@R: c(person(given = "Maxime",
            family = "Faymonville",
            role = c("aut", "cre"),
