+++ R/spinar_penal_val.R
@@ -44,7 +44,7 @@
 #' dat1 <- spinar_sim(n = 50, p = 1, alpha = 0.5,
 #'                    pmf = c(0.3, 0.3, 0.2, 0.1, 0.1))
 #'
-#' \dontrun{
+#' \donttest{
 #' # penalized semiparametric estimation with validation over L1
 #' spinar_penal_val(x = dat1, p = 1, validation = TRUE, penal2 = 0.1,
 #'                  over = "L1")
