+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: spINAR
 Type: Package
-Title: Semiparametric and Parametric Estimation and Bootstrapping of Integer-Valued Autoregressive (INAR) Models
+Title: (Semi)parametric Estimation and Bootstrapping of INAR Models
 Version: 0.1.0
 Date: 2023-03-22
 Authors@R: c(person(given = "Maxime",
