+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: spINAR
 Type: Package
-Title: (Semi)parametric Estimation and Bootstrapping of INAR Models
+Title: (Semi)Parametric Estimation and Bootstrapping of INAR Models
 Version: 0.1.0
 Date: 2023-03-22
 Authors@R: c(person(given = "Maxime",
@@ -24,10 +24,8 @@ Authors@R: c(person(given = "Maxime",
            email = "jentsch@statistik.tu-dortmund.de",
            comment = c(ORCID = "0000-0001-7824-1697")))
 Maintainer: Maxime Faymonville <faymonville@statistik.tu-dortmund.de>
-Description: Semiparametric and parametric estimation of INAR models including a finite sample refinement for the semiparametric setting, different procedures to bootstrap INAR data and flexible simulation of INAR data.
-		Faymonville, M., Jentsch, C., Weiß, C.H. and Aleksandrov, B. (2022) "Semiparametric Estimation of INAR Models using Roughness Penalization" <doi:10.1007/s10260-022-00655-0>
-		Jentsch, C. and Weiß, C.H. (2017) "Bootstrapping INAR Models" <doi:10.3150/18-BEJ1057>
-		Drost, F., Van den Akker, R. and Werker, B. (2009) "Efficient estimation of auto-regression parameters and innovation distributions for semiparametric integer-valued AR(p) models" <doi:10.1111/j.1467-9868.2008.00687.x>
+Description: Semiparametric and parametric estimation of INAR models including a finite sample refinement (Faymonville et al. (2022) <doi:10.1007/s10260-022-00655-0>) for the semiparametric setting introduced in Drost et al. (2009) 
+<doi:10.1111/j.1467-9868.2008.00687.x>, different procedures to bootstrap INAR data (Jentsch, C. and Weiß, C.H. (2017) <doi:10.3150/18-BEJ1057>) and flexible simulation of INAR data.
 License: GPL (>= 3)
 Encoding: UTF-8
 Depends:
