### 2023-04-19 MFaymon/spINAR#4 @ManuelStapper
Review
Hi there,

I reviewed your package/paper and genuinely enjoyed learning about the topic and testing the package.  There are a few comments and questions, I'd like to address regarding my review checklist:

- Scholarly effort: The package is probably flagged for inspection by the editors due to <1000 LOC (according to codecov), but I see the effort in structuring the package and the rather complex theory. However, I was wondering if there had been previous work you relied on? I.e  an implementation of methods by Carsten Jentsch and Christian Weiß from their 2019 paper on the Bootstrap procedure? Or am implementation by Drost et al.?

- Performance: You state (around line 48 of the paper) that your main contribution is the efficient semiparametric estimation. It is not clear to me whether "efficient" refers to the properties of the estimators or the implementation of estimation.

- Statement of Need: In my opinion the SoN in your paper and/or documentation would greatly benefit if you mentioned an application example. Did other authors maybe even discuss an application in which the semiparametric estimation is superior to a parametric approach (as that's your focus)?

-  Statement of Need / State of field: You mentioned that INAR processes are "clearly the most popular" among observation-driven count data models (around line 11 of the paper). On what basis do you make the statement? And related question: Is there actually no implementation in R dealing with INAR models, at least for parametric estimation? That would be in contrast with above popularity statement, wouldn't it? For example the tscount package has been applied and cited quite a few times.

- Community guidelines: JOSS suggests to give the target audience information on how to contribute and how to get support. I think it would be a great idea to include a note in your Repo.

- Outlook: I am very curious about how the package will develop in the future. If you have any plans, maybe you could list them in the Repo's readme. My suggestion would be to cover INAR(p) processes or arbitrary order by implementing a general likelihood function and leave the choice to the users.

Looking forward to hearing from you.
All the best and keep up the good work!
