### 2023-04-19 #5386 @ManuelStapper
Opened an issue over at the package Repo ([HERE](https://github.com/MFaymon/spINAR/issues/4)) for comments and questions.
