### 2023-05-03 MFaymon/spINAR#4 @MFaymon
> Community guidelines: JOSS suggests to give the target audience information on how to contribute and how to get support. I think it would be a great idea to include a note in your Repo.

Thanks for the suggestion. I included a note in the Readme, see 8283b92.
