### 2023-05-03 MFaymon/spINAR#4 @MFaymon
> Outlook: I am very curious about how the package will develop in the future. If you have any plans, maybe you could list them in the Repo's readme. My suggestion would be to cover INAR(p) processes or arbitrary order by implementing a general likelihood function and leave the choice to the users.

I added a possible extension to the Readme, see ea325f7 and dc3168c, which came to our mind during writing the package. It contains a generalization of the INAR model class.

We do not see much benefit in the inclusion of higher order INAR models ($p >2$) in our package since mostly only INAR models of order $p \in \{ 1,2 \}$ find application in the literature.
