### 2023-05-04 MFaymon/spINAR#4 @MFaymon
> Scholarly effort: The package is probably flagged for inspection by the editors due to <1000 LOC (according to codecov), but I see the effort in structuring the package and the rather complex theory. However, I was wondering if there had been previous work you relied on? I.e an implementation of methods by Carsten Jentsch and Christian Weiß from their 2019 paper on the Bootstrap procedure? Or am implementation by Drost et al.?

Thanks for bringing it up! We don‘t use any implementation of Drost et al. but we relied on some excerpt of code from Christian Weiß. After consultation with him, I added him as contributor of our package, see 6adef07.
