### 2023-05-04 MFaymon/spINAR#4 @MFaymon
> Performance: You state (around line 48 of the paper) that your main contribution is the efficient semiparametric estimation. It is not clear to me whether "efficient" refers to the properties of the estimators or the implementation of estimation.

Efficiency is a property of the semiparametric estimator, see [Drost et al. (2009)](https://doi.org/10.1111/j.1467-9868.2008.00687.x). To avoid misunderstandings, I have changed the wording in the paper, see 216657b.
