+++ paper/paper.bib
@@ -101,6 +101,13 @@ doi = {10.18637/jss.v082.i05},
   edition={1}
 }
 
+@book{dissweiss,
+  title     = {Categorical Times Series Analysis and Applications in Statistical Quality Control},
+  publisher = {dissertation.de},
+  year      = {2009},
+  author    = {Weiß, C.H.}
+}
+
 @electronic{rcoreteam,
   title     = {R: {A} {Language} and {Environment} for {Statistical} {Computing}},
   publisher = {R Foundation for Statistical Computing},
