+++ paper/paper.md
@@ -34,7 +34,7 @@ journal: JOSS
 
 # Summary
 
-While the statistical literature on continuous-valued time series processes is vast and the toolbox for parametric, non-parametric and semiparametric approaches is methodologically sound, the literature on count data time series is considerably less developed. Such count data time series models are usually categorized in parameter-driven and observation-driven models. Among the observation-driven approaches, the integer-valued autoregressive (INAR) models that rely on the famous binomial thinning operation due to @steutel are clearly the most popular ones. In particular, the INAR($p$) model can be seen as the discrete analogue of the well-known AR($p$) model for continuous-valued time series. The INAR(1) model was first introduced by @alosh and @mck, and its extension to the INAR($p$) model by Du and Li (1991) is defined according to $$X_t = \alpha_1 \circ X_{t-1} + \alpha_2 \circ X_{t-2} + \ldots + \alpha_p \circ X_{t-p} + \varepsilon_t, $$ with  $\varepsilon_t \overset{\text{i.i.d.}}{\sim} G$, where the innovation distribution $G$ has range $\mathbb{N}_0=\{0,1,2, \ldots\}$. The vector of INAR coefficients $\alpha = (\alpha_1, \ldots, \alpha_p)' \in (0,1)^p$ fulfills $\sum_{i=1}^p \alpha_i < 1$ and $$\alpha_i \circ X_{t-i} = \sum\limits_{j=1}^{X_{t-i}} Z_j^{(t,i)}, \, Z_j^{(t,i)} \sim \text{Bin}(1, \alpha_i), $$ where "$\circ$" denotes the binomial thinning operator first introduced by @steutel. Although many contributions have been made during the last decades, most of the literature focuses on parametric INAR models and estimation techniques. We want to emphasize the efficient semiparametric estimation of INAR models [@drost].
+While the statistical literature on continuous-valued time series processes is vast and the toolbox for parametric, non-parametric and semiparametric approaches is methodologically sound, the literature on count data time series is considerably less developed. Such count data time series models are usually categorized in parameter-driven and observation-driven models. Among the observation-driven approaches, the integer-valued autoregressive (INAR) models that rely on the famous binomial thinning operation due to @steutel are arguably the most popular ones. They have a simple intuitive and easy interpretable structure and have been widely applied in practice [@dissweiss]. In particular, the INAR($p$) model can be seen as the discrete analogue of the well-known AR($p$) model for continuous-valued time series. The INAR(1) model was first introduced by @alosh and @mck, and its extension to the INAR($p$) model by Du and Li (1991) is defined according to $$X_t = \alpha_1 \circ X_{t-1} + \alpha_2 \circ X_{t-2} + \ldots + \alpha_p \circ X_{t-p} + \varepsilon_t, $$ with  $\varepsilon_t \overset{\text{i.i.d.}}{\sim} G$, where the innovation distribution $G$ has range $\mathbb{N}_0=\{0,1,2, \ldots\}$. The vector of INAR coefficients $\alpha = (\alpha_1, \ldots, \alpha_p)' \in (0,1)^p$ fulfills $\sum_{i=1}^p \alpha_i < 1$ and $$\alpha_i \circ X_{t-i} = \sum\limits_{j=1}^{X_{t-i}} Z_j^{(t,i)}, \, Z_j^{(t,i)} \sim \text{Bin}(1, \alpha_i), $$ where "$\circ$" denotes the binomial thinning operator first introduced by @steutel. Although many contributions have been made during the last decades, most of the literature focuses on parametric INAR models and estimation techniques. We want to emphasize the efficient semiparametric estimation of INAR models [@drost].
 
 # Statement of need
 
