+++ paper/paper.bib
@@ -7,6 +7,13 @@ volume = {8(3)},
 pages = {261-275}
 }   
 
+@article{brae_stocks,
+author={Brännäs, K. and Shahiduzzaman, Q.},
+year = {2004},
+title={Integer-valued  moving average modelling of the number of transactions in stocks},
+journal = {Working Paper Economic Studies 637}
+}
+
 @article{drost,
 author={Drost, Feike and {Van den Akker}, Ramon and Werker, Bas},
 year = {2009},
@@ -24,6 +31,14 @@ journal = {Journal of Time Series Analysis},
 volume = {12(2)},
 pages = {129-142}
 }  
+
+@article{franke_med,
+author={Franke, J. and Seligmann, T.},
+year = {1993},
+title={Conditional maximum-likelihood estimates for INAR(1) processes and their applications to modeling epileptic seizure counts},
+journal = {Developments in Time Series},
+pages = {310-330}
+}
 
 @article{jewe,
 author={Jentsch, C. and Weiß, C.H.},
@@ -82,6 +97,15 @@ volume = {7(5)},
 pages = {893-899}
 }  
 
+@article{thy_rain,
+title = {Integer valued autoregressive models for tipping bucket rainfall measurements},
+author = {Thyregod, P. and Carstensen, J. and Madsen, H. and Arnbjerg-Nielsen, K.},
+journal = {Environmetrics},
+year = {1999},
+volume = {10},
+pages = {395-411}
+}
+
 @article{tscount,
 title = {{tscount}: An {R} Package for Analysis of Count Time Series Following Generalized Linear Models},
 author = {Liboschik, T. and Fokianos, K. and Fried, R.},
@@ -113,5 +137,14 @@ doi = {10.18637/jss.v082.i05},
   publisher = {R Foundation for Statistical Computing},
   year      = {2023},
   author    = {{R Core Team}},
-  url       = {https://www.R-project.org/},
+  url       = {https://www.R-project.org/}
 }
+
+@Manual{zinarp,
+  title     = {ZINARp: Simulate INAR/ZINAR(p) Models and Estimate Its Parameters},
+  author    = {Aldo William {Medina Garay} and Francyelle {de Lima Medina} and Tharso Augustus {Rossiter Araújo Monteiro}},
+  year      = {2022},
+  note = {R package version 0.1.0},
+  url = {https://CRAN.R-project.org/package=ZINARp}
+}
+
