+++ paper/paper.md
@@ -38,7 +38,7 @@ While the statistical literature on continuous-valued time series processes is v
 
 # Statement of need
 
-To deal with count data time series, [`R`](https://www.r-project.org/) [@rcoreteam] provides the package [`tscount`](https://cran.r-project.org/web/packages/tscount/) [@tscount] which, a.o., includes likelihood-based estimation of parameter-driven count data time series models which do not include INAR models and exclusively allows for conditional Poisson or negative binomially distributed data. The [`Julia`](https://julialang.org/) [@julia] package [`CountTimeSeries`](https://zenodo.org/record/7488440#.Y9ky9ISZNaQ) [@manuel] deals with integer counterparts of ARMA and GARCH models and some generalizations including the INAR model. While this package covers parametric estimation of INAR models, that is, of the INAR coefficients and of a parametrically specified innovation distribution $\{G_\theta \,  | \, \theta \in \mathbb{R}^q, \, q \in \mathbb{N}\}$, it does not allow for non-parametric estimation of the innovation distribution. Such a semiparametric estimation technique that still relies on the binomial thinning operation, but comes along without any parametric specification of the innovation distribution was proposed and proven to be efficient by @drost. Also neither of the two packages contains procedures for bootstrapping INAR models within these parametric and semiparametric setups. The [`R`](https://www.r-project.org/) package [`spINAR`](https://github.com/MFaymon/spINAR) fills this gap and combines simulation, estimation and bootstrapping of INAR models in a single package. Both, the estimation and the bootstrapping, are implemented semiparametrically and also parametrically. The package covers INAR models of order $p \in \{1,2\}$, which are mainly used in applications.
+To deal with count data time series, [`R`](https://www.r-project.org/) [@rcoreteam] provides the package [`tscount`](https://cran.r-project.org/web/packages/tscount/) [@tscount] which, a.o., includes likelihood-based estimation of parameter-driven count data time series models which do not include INAR models and exclusively allows for conditional Poisson or negative binomially distributed data. The R package [`ZINARp`](https://CRAN.R-project.org/package=ZINARp) [@zinarp] allows to simulate and estimate INAR data by using MCMC algorithms for estimation but the package is limited to parametric estimation of INAR models, that is, of the INAR coefficients and of a parametrically specified innovation distribution $\{G_\theta \,  | \, \theta \in \mathbb{R}^q, \, q \in \mathbb{N}\}$ where they only cover the cases of Poisson or zero-inflated Poisson distributed innovations.The [`Julia`](https://julialang.org/) [@julia] package [`CountTimeSeries`](https://zenodo.org/record/7488440#.Y9ky9ISZNaQ) [@manuel] deals with integer counterparts of ARMA and GARCH models and some generalizations including the INAR model. It covers the parametric estimation setup for INAR models but does also not allow for non-parametric estimation of the innovation distribution. Such a semiparametric estimation technique that still relies on the binomial thinning operation, but comes along without any parametric specification of the innovation distribution was proposed and proven to be efficient by @drost. Also neither of the three packages contains procedures for bootstrapping INAR models within these parametric and semiparametric setups. The [`R`](https://www.r-project.org/) package [`spINAR`](https://github.com/MFaymon/spINAR) fills this gap and combines simulation, estimation and bootstrapping of INAR models in a single package. Both, the estimation and the bootstrapping, are implemented semiparametrically and also parametrically. The package covers INAR models of order $p \in \{1,2\}$, which are mainly used in applications.
 
 # Features
 
