### 2023-05-12 MFaymon/spINAR#4 @MFaymon
> Statement of Need / State of field: You mentioned that INAR processes are "clearly the most popular" among observation-driven count data models (around line 11 of the paper). On what basis do you make the statement? And related question: Is there actually no implementation in R dealing with INAR models, at least for parametric estimation? That would be in contrast with above popularity statement, wouldn't it? For example the tscount package has been applied and cited quite a few times.

Thanks for pointing out that a source is missing there. I have weakened the statement a little and added a source citation, see 4e1a153.

As already mentioned, the R package tscount only deals with parameter-driven count data time series models. INAR models don’t belong to this model class. But I found a recent R packages which allows to simulate and estimate INAR(p) data by using MCMC algorithms. The package is limited to the parametric setup, where the innovations are assumed to follow either a Poisson or a zero-inflated Poisson distribution. I now cite it in the paper, see 190e847.
