+++ paper/paper.bib
@@ -7,13 +7,6 @@ volume = {8(3)},
 pages = {261-275}
 }   
 
-@article{brae_stocks,
-author={Brännäs, K. and Shahiduzzaman, Q.},
-year = {2004},
-title={Integer-valued  moving average modelling of the number of transactions in stocks},
-journal = {Working Paper Economic Studies 637}
-}
-
 @article{drost,
 author={Drost, Feike and {Van den Akker}, Ramon and Werker, Bas},
 year = {2009},
@@ -38,6 +31,15 @@ year = {1993},
 title={Conditional maximum-likelihood estimates for INAR(1) processes and their applications to modeling epileptic seizure counts},
 journal = {Developments in Time Series},
 pages = {310-330}
+}
+
+@article{jazi,
+author={Jazi,M. and Jones, G. and Lai, C.},
+year = {2012},
+title={First-order integer valued AR processes with zero inflated Poisson innovations},
+journal = {Journal of Time Series Analysis},
+volume = {33},
+pages = {954-963}
 }
 
 @article{jewe,
@@ -86,7 +88,16 @@ title={Some simple models for discrete variate time series},
 journal = {Water Resources Bulletin},
 volume = {21(4)},
 pages = {645-650}
-}     
+}   
+
+@article{mc_wage,
+author={McCabe, B. and Martin, G.},
+year = {2005},
+title={Bayesian predictions of low count time series},
+journal = {International Journal of Forecasting},
+volume = {21(2)},
+pages = {315-330}
+}   
 
 @article{steutel,
 author={Steutel, F. W. and Van Harn, K.},
