+++ paper/paper.md
@@ -38,6 +38,8 @@ While the statistical literature on continuous-valued time series processes is v
 
 # Statement of need
 
+INAR models find applications in a wide variety of fields such as medical sciences, environmentology and economics. For example, @franke_med model epileptic seizure counts using an INAR(1) model, @thy_rain use integer-valued autoregressive models to model the dynamics of rainfall and @mc_wage to analyze wage loss claims data. They all have in common assuming that the innovation distribution belongs to a parametric class of distributions. Non- or semiparametric estimation of the INAR model was not considered until @drost came up with their semiparametric estimation approach. A possible explanation is the complexity of the semiparametric setup since despite in the AR case the estimation in the INAR case cannot be based on the residuals: Even if the autoregressive coefficients were known, observing the data does not imply observing the innovations [@drost]. Nonetheless, one big advantage of semiparametric estimation is that we do not need to make a parametric distribution assumption on the innovations. The Poisson assumption is, for example, the most frequently used assumption for innovations and is characterized by equidispersion. In most cases, however, the data shows a higher variance than the mean value. The question arises when the distance between these two moments is large enough to not rather assume overdispersion, which would probably lead to assume negative binomially or geometrically distributed innovations. Furthermore, when dealing with low counts, we often observe many zeros in the data. This could be a sign for a zero-inflated innovation distribution such as the zero-inflated Poisson distribution [@jazi]. However, it is unclear at what point the zero is represented frequently enough in the data set to justify such an assumption. The mentioned points indicate that the assumption of an appropriate innovation distribution is often critical, bearing in mind that an incorrect assumption can lead to poor estimation performance. With semiparametric estimation, we do not have to commit to an innovation distribution, which makes this approach appealing. 
+
 To deal with count data time series, [`R`](https://www.r-project.org/) [@rcoreteam] provides the package [`tscount`](https://cran.r-project.org/web/packages/tscount/) [@tscount] which, a.o., includes likelihood-based estimation of parameter-driven count data time series models which do not include INAR models and exclusively allows for conditional Poisson or negative binomially distributed data. The R package [`ZINARp`](https://CRAN.R-project.org/package=ZINARp) [@zinarp] allows to simulate and estimate INAR data by using MCMC algorithms for estimation but the package is limited to parametric estimation of INAR models, that is, of the INAR coefficients and of a parametrically specified innovation distribution $\{G_\theta \,  | \, \theta \in \mathbb{R}^q, \, q \in \mathbb{N}\}$ where they only cover the cases of Poisson or zero-inflated Poisson distributed innovations.The [`Julia`](https://julialang.org/) [@julia] package [`CountTimeSeries`](https://zenodo.org/record/7488440#.Y9ky9ISZNaQ) [@manuel] deals with integer counterparts of ARMA and GARCH models and some generalizations including the INAR model. It covers the parametric estimation setup for INAR models but does also not allow for non-parametric estimation of the innovation distribution. Such a semiparametric estimation technique that still relies on the binomial thinning operation, but comes along without any parametric specification of the innovation distribution was proposed and proven to be efficient by @drost. Also neither of the three packages contains procedures for bootstrapping INAR models within these parametric and semiparametric setups. The [`R`](https://www.r-project.org/) package [`spINAR`](https://github.com/MFaymon/spINAR) fills this gap and combines simulation, estimation and bootstrapping of INAR models in a single package. Both, the estimation and the bootstrapping, are implemented semiparametrically and also parametrically. The package covers INAR models of order $p \in \{1,2\}$, which are mainly used in applications.
 
 # Features
