### 2023-05-12 MFaymon/spINAR#4 @MFaymon
> Statement of Need: In my opinion the SoN in your paper and/or documentation would greatly benefit if you mentioned an application example. Did other authors maybe even discuss an application in which the semiparametric estimation is superior to a parametric approach (as that's your focus)?

Thanks for this valuable comment! The paper now mentions several application examples and I added a discussion which emphasizes the benefit of the semiparametric estimation approach, see 62eefde.
