### 2023-09-08 #5386 @MFaymon
I suggest Dr. Philipp Wittenberg from Helmut-Schmidt-University Hamburg as reviewer. Part of his research deals with autoregressive counts. I suppose that @wittenberg is his GitHub profile.
