### 2024-02-29 40fcca6
add vignette with reproduced results from the literature
