### 2024-03-01 71f646e
update spinar_est_param, spinar_penal, spinar_penal_val
