### 2024-03-01 a9de6f3
update spinar_sim, spinar_boot and help page of spinar_boot
