### 2024-03-01 MFaymon/spINAR#5 @MFaymon
> 2\. Regarding the software package functionality, I performed tests on the provided functions and found that for some parameters (settings might be artificial), a check needs to be included to intercept these incorrect entries.

Thanks for pointing out that some checks were missing! The respective functions now include checks for interception of these incorrect inputs, see 4691b30, 71f646e and a9de6f3.
