### 2024-03-01 #5386 @MFaymon
Thanks a lot for the review @wittenberg! I adressed all your suggestions for improvement, see (https://github.com/MFaymon/spINAR/issues/5).
