### 2024-03-01 2901bce
added tests for spinar_sim
