### 2024-03-01 cf4399a
adjusted error message in spinar_sim
