### 2024-03-19 7ee951a
removed typos and vignettes added vignette as rmd
