+++ R/spINAR-package.R
@@ -20,7 +20,7 @@
 #'
 #' @references
 #' Faymonville, M., Jentsch, C., Weiß, C.H. and Aleksandrov, B. (2022).
-#' Semiparametric Estimation of INAR Models using Roughness Penalization.
+#' "Semiparametric Estimation of INAR Models using Roughness Penalization".
 #' Statistical Methods & Applications. \doi{10.1007/s10260-022-00655-0}.
 #'
 #' Jentsch, C. and Weiß, C. H. (2017), “Bootstrapping INAR Models”.
