### 2024-03-19 ce15355
added vignette of literature results as rmd
