+++ vignettes/Replicate_literature_results_using_spINAR.Rmd
@@ -198,7 +198,7 @@ est_penal <- spinar_penal(data,1,0,1)
 
 Plot of the unpenalized and penalized estimated innovation distribution
 
-```{r}
+```{r eval=FLASE}
 par(mfrow=c(1,2))
 barplot(est_unpenal[-1], ylim=c(0,1),names.arg=0:5,
 main="Unpenalized estimated \n innovation distribution")
@@ -206,7 +206,7 @@ barplot(est_penal[-1], ylim=c(0,1),names.arg=0:5,
 main="Penalized estimated \n innovation distribution")
 ```
 
-![](https://github.com/MFaymon/spINAR/blob/main/vignette/barplots.png)
+![](https://github.com/MFaymon/spINAR/blob/main/vignettes/barplots.png)
 ## References
 
 * Faymonville, M., Jentsch, C., Weiß, C.H. and Aleksandrov, B. (2022). "Semiparametric Estimation of INAR Models using Roughness Penalization". Statistical Methods & Applications. [DOI](https://doi.org/10.1007/s10260-022-00655-0)
