### 2024-04-02 e9dfcbd
removed typo in spinar_boot and its help page
