### 2024-04-02 54a303a
added condition in spinar_boot
