### 2024-04-02 bf10e41
update spinar_est_param
