### 2024-04-08 #5386 @MFaymon
Version number of latest submission to CRAN: v0.2.0
DOI of latest Zenodo release: [10.5281/zenodo.10944202]( https://doi.org/10.5281/zenodo.10944202)

The latest CRAN release is not published yet but submitted and on it's way. We also checked the other author's tasks, so there should be nothing left to do from our side :)
