### 2024-05-06 #5386 @crvernon
:wave: @MFaymon - I am the AEiC for the track this submission in on.  I am going to team with @crsl4 to move this one forward as you are very close to publication!

The following are a few minor things for you to address before we can move forward:

- [ ] The title and authors of your Zenodo record (https://zenodo.org/records/10944202) should match the title from your paper.  Please endure the authors also appear in the same order if they do not already.  Please also change the license specified in your Zenodo archive from CC 4.0 to what is in your repo for your software (GPL-3.0).  You can do all of this without having to make another release...just edit the metadata for the record and you should be good.
- [ ]  Please check the following references for potential DOIs as rendered from https://github.com/openjournals/joss-reviews/issues/5386#issuecomment-2096219998:

> - 10.1111/j.1467-9892.2012.00809.x may be a valid DOI for title: First-order integer valued AR processes with zero ...
> - 10.3150/18-bej1057 may be a valid DOI for title: Bootstrapping INAR Models
> - 10.1016/j.ijforecast.2004.11.001 may be a valid DOI for title: Bayesian predictions of low count time series
> - 10.1214/aop/1176994950 may be a valid DOI for title: Discrete analogues of self-decomposability and sta...

There is no reason to make any new releases when incorporating any changes above since these are all in reference to the paper or archive.

Thanks so much and I look forward to hearing back from you soon so we can get all of your hard work in the books!
