+++ paper/paper.bib
@@ -39,7 +39,8 @@ year = {2012},
 title={First-order integer valued AR processes with zero inflated Poisson innovations},
 journal = {Journal of Time Series Analysis},
 volume = {33},
-pages = {954-963}
+pages = {954-963},
+doi={10.1111/j.1467-9892.2012.00809.x}
 }
 
 @article{jewe,
@@ -48,7 +49,8 @@ year = {2017},
 title={Bootstrapping INAR Models},
 journal = {Bernoulli},
 volume = {25(3)},
-pages = {2359-2408}
+pages = {2359-2408},
+doi={10.3150/18-BEJ1057}
 }
 
 @article{julia,
@@ -96,7 +98,8 @@ year = {2005},
 title={Bayesian predictions of low count time series},
 journal = {International Journal of Forecasting},
 volume = {21(2)},
-pages = {315-330}
+pages = {315-330},
+doi={10.1016/j.ijforecast.2004.11.001}
 }   
 
 @article{steutel,
@@ -105,7 +108,8 @@ year = {1979},
 title={Discrete analogues of self-decomposability and stability},
 journal = {Annals of Probability},
 volume = {7(5)},
-pages = {893-899}
+pages = {893-899},
+doi={10.1214/aop/1176994950}
 }  
 
 @article{thy_rain,
