### 2024-05-07 #5386 @crvernon
@MFaymon could you do a last check of the proof that was just generated [here](https://github.com/openjournals/joss-reviews/issues/5386#issuecomment-2098803689).  Once you confirm, I'll push it to publish.
