### 2024-05-08 #5386 @crvernon
Congratulations on your new publication @MFaymon! Many thanks to @crsl4 for editing and @ManuelStapper and @wittenberg for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
